requirements.txt
fastapi>=0.110.0
slowapi>=0.1.9
pycryptodome>=3.20.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
httpx>=0.26.0
websockets>=12.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
sse-starlette>=2.0.0
sydney-py>=0.23.0
openai>=1.10.0
python-multipart>=0.0.6
Pillow>=10.0.0
requests>=2.31.0
cachetools>=5.3.0
uvloop>=0.19.0
orjson>=3.9.0
tiktoken>=0.7.0

# Dev / testing dependencies
pytest>=8.0
pytest-asyncio>=0.23
aioresponses>=0.7.6